test(js): move sign-up prepare coalescing test into SignUp suite
